django/requirements/requirements_dev.txt
asgiref==3.7.2
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.3.2
coreapi==2.3.3
coreschema==0.0.4
cryptography==41.0.7
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
dj-rest-auth==5.0.2
drf-yasg==1.21.7
idna==3.6
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.2
MarkupSafe==2.1.3
packaging==23.2
pycparser==2.21
pyparsing==3.1.1
pytz==2023.3.post1
requests==2.31.0
ruamel.yaml==0.18.5
ruamel.yaml.clib==0.2.8
sqlparse==0.4.4
tzdata==2023.3
uritemplate==4.1.1
urllib3==2.1.0
channels==4.0.0
daphne==4.0.0
pymunk==6.6.0
